service object created and working with a test
