--- app/services/news_api_service.rb
@@ -1,16 +1,0 @@
-require 'news-api'
-
-class NewsApiService
-
-  def initialize
-    @api_key = "3222ed90cb3844f68f39c38c28927850"
-    # need to set api_key through ENV though
-  end
-
-  # all the business logic will go here
-  # NewsApiServices.new.search will later be called in controllers
-
-  def search
-  end
-
-end
